+++ stdlib/LinearAlgebra/src/LinearAlgebra.jl
@@ -138,6 +138,8 @@ export
     svdvals!,
     svdvals,
     sylvester,
+    symmetrize,
+    symmetrize!,
     tr,
     transpose,
     transpose!,